Properties and Types
Added a whole bunch of new property and property types.
Properties now have a `$confirmance` array.
`CalendarObject` will now now do a `conformanceTest` on all properties and will throw a `FailedConformanceTestException` if the property cannot be added to the calendar object.
The `Parameter` class now has a `getParameter` method which is called by `__toString()`. The public `toString()` method will only return the value, not the whole parameter string.
There's a new `MultipleValues` trait which can be added to a propery to allow multiple values.